testing/web-platform/tests/fetch/corb/script-html-js-polyglot.sub.html

<!DOCTYPE html>
<!-- Test verifies that CORB won't block a polyglot script that is
     both a valid HTML document and also valid Javascript.
-->
<meta charset="utf-8">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var script = document.createElement("script")

  script.onload = t.step_func_done(function(){
    // Verify that html-js-polyglot.js wasn't blocked - that script
    // should have set window.polyglot to 123.
    assert_equals(window.polyglot, 123);
  })
  addEventListener("error",function(e) {
    t.step(function() {
      assert_unreached("No errors are expected with or without CORB.");
      t.done();
    })
  });

  // www1 is cross-origin, so the HTTP response is CORB-eligible.
  script.src = "http://{{domains[www1]}}:{{ports[http][0]}}/fetch/corb/resources/html-js-polyglot.js"
  document.body.appendChild(script)
}, "CORB cannot block polyglot HTML/JS");
</script>
